+++ DateTime.js
@@ -1,25 +1,15 @@
 'use strict';
 
 var assign = require('object-assign'),
+        PropTypes = require('prop-types'),
+        createClass = require('create-react-class'),
+	moment = require('moment'),
 	React = require('react'),
-	DaysView = require('./src/DaysView'),
-	MonthsView = require('./src/MonthsView'),
-	YearsView = require('./src/YearsView'),
-	TimeView = require('./src/TimeView'),
-	moment = require('moment')
+	CalendarContainer = require('./src/CalendarContainer')
 ;
 
-var TYPES = React.PropTypes;
-var Datetime = React.createClass({
-	mixins: [
-		require('./src/onClickOutside')
-	],
-	viewComponents: {
-		days: DaysView,
-		months: MonthsView,
-		years: YearsView,
-		time: TimeView
-	},
+var TYPES = PropTypes;
+var Datetime = createClass({
 	propTypes: {
 		// value: TYPES.object | TYPES.string,
 		// defaultValue: TYPES.object | TYPES.string,
@@ -42,7 +32,7 @@ var Datetime = React.createClass({
 	},
 
 	getDefaultProps: function() {
-		var nof = function(){};
+		var nof = function() {};
 		return {
 			className: '',
 			defaultValue: '',
@@ -72,7 +62,7 @@ var Datetime = React.createClass({
 		return state;
 	},
 
-	getStateFromProps: function( props ){
+	getStateFromProps: function( props ) {
 		var formats = this.getFormats( props ),
 			date = props.value || props.defaultValue,
 			selectedDate, viewDate, updateOn, inputValue
@@ -110,36 +100,36 @@ var Datetime = React.createClass({
 		};
 	},
 
-	getUpdateOn: function(formats){
-		if ( formats.date.match(/[lLD]/) ){
+	getUpdateOn: function( formats ) {
+		if ( formats.date.match(/[lLD]/) ) {
 			return 'days';
 		}
-		else if ( formats.date.indexOf('M') !== -1 ){
+		else if ( formats.date.indexOf('M') !== -1 ) {
 			return 'months';
 		}
-		else if ( formats.date.indexOf('Y') !== -1 ){
+		else if ( formats.date.indexOf('Y') !== -1 ) {
 			return 'years';
 		}
 
 		return 'days';
 	},
 
-	getFormats: function( props ){
+	getFormats: function( props ) {
 		var formats = {
 				date: props.dateFormat || '',
 				time: props.timeFormat || ''
 			},
-			locale = this.localMoment( props.date ).localeData()
+			locale = this.localMoment( props.date, null, props ).localeData()
 		;
 
-		if ( formats.date === true ){
+		if ( formats.date === true ) {
 			formats.date = locale.longDateFormat('L');
 		}
-		else if ( this.getUpdateOn(formats) !== 'days' ){
+		else if ( this.getUpdateOn(formats) !== 'days' ) {
 			formats.time = '';
 		}
 
-		if ( formats.time === true ){
+		if ( formats.time === true ) {
 			formats.time = locale.longDateFormat('LT');
 		}
 
@@ -151,26 +141,59 @@ var Datetime = React.createClass({
 		return formats;
 	},
 
-	componentWillReceiveProps: function(nextProps) {
+	componentWillReceiveProps: function( nextProps ) {
 		var formats = this.getFormats( nextProps ),
-			update = {}
+			updatedState = {}
 		;
 
 		if ( nextProps.value !== this.props.value ||
-            formats.datetime !== this.getFormats( this.props ).datetime ){
-            update = this.getStateFromProps( nextProps );
+			formats.datetime !== this.getFormats( this.props ).datetime ) {
+			updatedState = this.getStateFromProps( nextProps );
 		}
 
-		if ( update.open === undefined ){
-			if ( this.props.closeOnSelect && this.state.currentView !== 'time' ){
-				update.open = false;
-			}
-			else {
-				update.open = this.state.open;
+		if ( updatedState.open === undefined ) {
+			if ( this.props.closeOnSelect && this.state.currentView !== 'time' ) {
+				updatedState.open = false;
+			} else {
+				updatedState.open = this.state.open;
 			}
 		}
 
-		this.setState( update );
+		if ( nextProps.viewMode !== this.props.viewMode ) {
+			updatedState.currentView = nextProps.viewMode;
+		}
+
+		if ( nextProps.locale !== this.props.locale ) {
+			if ( this.state.viewDate ) {
+				var updatedViewDate = this.state.viewDate.clone().locale( nextProps.locale );
+				updatedState.viewDate = updatedViewDate;
+			}
+			if ( this.state.selectedDate ) {
+				var updatedSelectedDate = this.state.selectedDate.clone().locale( nextProps.locale );
+				updatedState.selectedDate = updatedSelectedDate;
+				updatedState.inputValue = updatedSelectedDate.format( formats.datetime );
+			}
+		}
+
+		if ( nextProps.utc !== this.props.utc ) {
+			if ( nextProps.utc ) {
+				if ( this.state.viewDate )
+					updatedState.viewDate = this.state.viewDate.clone().utc();
+				if ( this.state.selectedDate ) {
+					updatedState.selectedDate = this.state.selectedDate.clone().utc();
+					updatedState.inputValue = updatedState.selectedDate.format( formats.datetime );
+				}
+			} else {
+				if ( this.state.viewDate )
+					updatedState.viewDate = this.state.viewDate.clone().local();
+				if ( this.state.selectedDate ) {
+					updatedState.selectedDate = this.state.selectedDate.clone().local();
+					updatedState.inputValue = updatedState.selectedDate.format(formats.datetime);
+				}
+			}
+		}
+
+		this.setState( updatedState );
 	},
 
 	onInputChange: function( e ) {
@@ -192,27 +215,27 @@ var Datetime = React.createClass({
 		});
 	},
 
-	onInputKey: function( e ){
-		if ( e.which === 9 && this.props.closeOnTab ){
+	onInputKey: function( e ) {
+		if ( e.which === 9 && this.props.closeOnTab ) {
 			this.closeCalendar();
 		}
 	},
 
-	showView: function( view ){
+	showView: function( view ) {
 		var me = this;
-		return function(){
+		return function() {
 			me.setState({ currentView: view });
 		};
 	},
 
-	setDate: function( type ){
+	setDate: function( type ) {
 		var me = this,
 			nextViews = {
 				month: 'days',
 				year: 'months'
 			}
 		;
-		return function( e ){
+		return function( e ) {
 			me.setState({
 				viewDate: me.state.viewDate.clone()[ type ]( parseInt(e.target.getAttribute('data-value'), 10) ).startOf( type ),
 				currentView: nextViews[ type ]
@@ -220,18 +243,18 @@ var Datetime = React.createClass({
 		};
 	},
 
-	addTime: function( amount, type, toSelected ){
+	addTime: function( amount, type, toSelected ) {
 		return this.updateTime( 'add', amount, type, toSelected );
 	},
 
-	subtractTime: function( amount, type, toSelected ){
+	subtractTime: function( amount, type, toSelected ) {
 		return this.updateTime( 'subtract', amount, type, toSelected );
 	},
 
-	updateTime: function( op, amount, type, toSelected ){
+	updateTime: function( op, amount, type, toSelected ) {
 		var me = this;
 
-		return function(){
+		return function() {
 			var update = {},
 				date = toSelected ? 'selectedDate' : 'viewDate'
 			;
@@ -243,7 +266,7 @@ var Datetime = React.createClass({
 	},
 
 	allowedSetTime: ['hours', 'minutes', 'seconds', 'milliseconds'],
-	setTime: function( type, value ){
+	setTime: function( type, value ) {
 		var index = this.allowedSetTime.indexOf( type ) + 1,
 			state = this.state,
 			date = (state.selectedDate || state.viewDate).clone(),
@@ -258,7 +281,7 @@ var Datetime = React.createClass({
 			date[ nextType ]( date[nextType]() );
 		}
 
-		if ( !this.props.value ){
+		if ( !this.props.value ) {
 			this.setState({
 				selectedDate: date,
 				inputValue: date.format( state.inputFormat )
@@ -275,7 +298,7 @@ var Datetime = React.createClass({
 			date
     ;
 
-		if (target.className.indexOf('rdtDay') !== -1){
+		if (target.className.indexOf('rdtDay') !== -1) {
 			if (target.className.indexOf('rdtNew') !== -1)
 				modifier = 1;
 			else if (target.className.indexOf('rdtOld') !== -1)
@@ -284,11 +307,11 @@ var Datetime = React.createClass({
 			date = viewDate.clone()
 				.month( viewDate.month() + modifier )
 				.date( parseInt( target.getAttribute('data-value'), 10 ) );
-		} else if (target.className.indexOf('rdtMonth') !== -1){
+		} else if (target.className.indexOf('rdtMonth') !== -1) {
 			date = viewDate.clone()
 				.month( parseInt( target.getAttribute('data-value'), 10 ) )
 				.date( currentDate.date() );
-		} else if (target.className.indexOf('rdtYear') !== -1){
+		} else if (target.className.indexOf('rdtYear') !== -1) {
 			date = viewDate.clone()
 				.month( currentDate.month() )
 				.date( currentDate.date() )
@@ -300,15 +323,20 @@ var Datetime = React.createClass({
 			.seconds( currentDate.seconds() )
 			.milliseconds( currentDate.milliseconds() );
 
-		if ( !this.props.value ){
+		if ( !this.props.value ) {
+			var open = !( this.props.closeOnSelect && close );
+			if ( !open ) {
+				this.props.onBlur( date );
+			}
+
 			this.setState({
 				selectedDate: date,
 				viewDate: date.clone().startOf('month'),
 				inputValue: date.format( this.state.inputFormat ),
-				open: !(this.props.closeOnSelect && close )
+				open: open
 			});
 		} else {
-			if (this.props.closeOnSelect && close) {
+			if ( this.props.closeOnSelect && close ) {
 				this.closeCalendar();
 			}
 		}
@@ -330,41 +358,42 @@ var Datetime = React.createClass({
 		});
 	},
 
-	handleClickOutside: function(){
-		if ( this.props.input && this.state.open && !this.props.open ){
+	handleClickOutside: function() {
+		if ( this.props.input && this.state.open && !this.props.open ) {
 			this.setState({ open: false }, function() {
 				this.props.onBlur( this.state.selectedDate || this.state.inputValue );
 			});
 		}
 	},
 
-	localMoment: function( date, format ){
-		var momentFn = this.props.utc ? moment.utc : moment;
-		var m = momentFn( date, format, this.props.strictParsing );
-		if ( this.props.locale )
-			m.locale( this.props.locale );
+	localMoment: function( date, format, props ) {
+		props = props || this.props;
+		var momentFn = props.utc ? moment.utc : moment;
+		var m = momentFn( date, format, props.strictParsing );
+		if ( props.locale )
+			m.locale( props.locale );
 		return m;
 	},
 
 	componentProps: {
 		fromProps: ['value', 'isValidDate', 'renderDay', 'renderMonth', 'renderYear', 'timeConstraints'],
 		fromState: ['viewDate', 'selectedDate', 'updateOn'],
-		fromThis: ['setDate', 'setTime', 'showView', 'addTime', 'subtractTime', 'updateSelectedDate', 'localMoment']
+		fromThis: ['setDate', 'setTime', 'showView', 'addTime', 'subtractTime', 'updateSelectedDate', 'localMoment', 'handleClickOutside']
 	},
 
-	getComponentProps: function(){
+	getComponentProps: function() {
 		var me = this,
 			formats = this.getFormats( this.props ),
 			props = {dateFormat: formats.date, timeFormat: formats.time}
 		;
 
-		this.componentProps.fromProps.forEach( function( name ){
+		this.componentProps.fromProps.forEach( function( name ) {
 			props[ name ] = me.props[ name ];
 		});
-		this.componentProps.fromState.forEach( function( name ){
+		this.componentProps.fromState.forEach( function( name ) {
 			props[ name ] = me.state[ name ];
 		});
-		this.componentProps.fromThis.forEach( function( name ){
+		this.componentProps.fromThis.forEach( function( name ) {
 			props[ name ] = me[ name ];
 		});
 
@@ -372,18 +401,17 @@ var Datetime = React.createClass({
 	},
 
 	render: function() {
-		var Component = this.viewComponents[ this.state.currentView ],
-			DOM = React.DOM,
+		var DOM = React.DOM,
 			className = 'rdt' + (this.props.className ?
                   ( Array.isArray( this.props.className ) ?
                   ' ' + this.props.className.join( ' ' ) : ' ' + this.props.className) : ''),
 			children = []
 		;
 
-		if ( this.props.input ){
+		if ( this.props.input ) {
 			children = [ DOM.input( assign({
 				key: 'i',
-				type:'text',
+				type: 'text',
 				className: 'form-control',
 				onFocus: this.openCalendar,
 				onChange: this.onInputChange,
@@ -400,7 +428,7 @@ var Datetime = React.createClass({
 		return DOM.div({className: className}, children.concat(
 			DOM.div(
 				{ key: 'dt', className: 'rdtPicker' },
-				React.createElement( Component, this.getComponentProps())
+				React.createElement( CalendarContainer, {view: this.state.currentView, viewProps: this.getComponentProps(), onClickOutside: this.handleClickOutside })
 			)
 		));
 	}