+++ src/MonthsView.js
@@ -1,17 +1,20 @@
 'use strict';
 
-var React = require('react');
+var React = require('react'),
+    createClass = require('create-react-class'),
+	onClickOutside = require('react-onclickoutside')
+;
 
 var DOM = React.DOM;
-var DateTimePickerMonths = React.createClass({
+var DateTimePickerMonths = onClickOutside( createClass({
 	render: function() {
 		return DOM.div({ className: 'rdtMonths' }, [
-			DOM.table({ key: 'a'}, DOM.thead({}, DOM.tr({}, [
-				DOM.th({ key: 'prev', className: 'rdtPrev' }, DOM.span({onClick: this.props.subtractTime(1, 'years')}, '‹')),
-				DOM.th({ key: 'year', className: 'rdtSwitch', onClick: this.props.showView('years'), colSpan: 2, 'data-value': this.props.viewDate.year()}, this.props.viewDate.year() ),
-				DOM.th({ key: 'next', className: 'rdtNext' }, DOM.span({onClick: this.props.addTime(1, 'years')}, '›'))
+			DOM.table({ key: 'a' }, DOM.thead( {}, DOM.tr( {}, [
+				DOM.th({ key: 'prev', className: 'rdtPrev', onClick: this.props.subtractTime( 1, 'years' )}, DOM.span({}, '‹' )),
+				DOM.th({ key: 'year', className: 'rdtSwitch', onClick: this.props.showView( 'years' ), colSpan: 2, 'data-value': this.props.viewDate.year() }, this.props.viewDate.year() ),
+				DOM.th({ key: 'next', className: 'rdtNext', onClick: this.props.addTime( 1, 'years' )}, DOM.span({}, '›' ))
 			]))),
-			DOM.table({ key: 'months'}, DOM.tbody({ key: 'b'}, this.renderMonths()))
+			DOM.table({ key: 'months' }, DOM.tbody({ key: 'b' }, this.renderMonths()))
 		]);
 	},
 
@@ -23,23 +26,33 @@ var DateTimePickerMonths = React.createClass({
 			i = 0,
 			months = [],
 			renderer = this.props.renderMonth || this.renderMonth,
-			isValid = this.props.isValidDate || this.isValidDate,
-			classes, props
+			isValid = this.props.isValidDate || this.alwaysValidDate,
+			classes, props, currentMonth, isDisabled, noOfDaysInMonth, daysInMonth, validDay,
+			// Date is irrelevant because we're only interested in month
+			irrelevantDate = 1
 		;
 
-        var currentMonth, disabled,
-            // Date is irrelevant because we're really only interested in month
-            irrelevantDate = 1;
 		while (i < 12) {
 			classes = 'rdtMonth';
 			currentMonth =
-                this.props.viewDate.clone().set({ year: year, month: i, date: irrelevantDate });
-			disabled = !isValid(currentMonth);
+				this.props.viewDate.clone().set({ year: year, month: i, date: irrelevantDate });
 
-			if ( disabled )
+			noOfDaysInMonth = currentMonth.endOf( 'month' ).format( 'D' );
+			daysInMonth = Array.from({ length: noOfDaysInMonth }, function( e, i ) {
+				return i + 1;
+			});
+
+			validDay = daysInMonth.find(function( d ) {
+				var day = currentMonth.clone().set( 'date', d );
+				return isValid( day );
+			});
+
+			isDisabled = ( validDay === undefined );
+
+			if ( isDisabled )
 				classes += ' rdtDisabled';
 
-			if ( date && i === month && year === date.year() )
+			if ( date && i === date.month() && year === date.year() )
 				classes += ' rdtActive';
 
 			props = {
@@ -48,14 +61,14 @@ var DateTimePickerMonths = React.createClass({
 				className: classes
 			};
 
-			if ( !disabled )
-				props.onClick = (this.props.updateOn === 'months' ?
-                    this.updateSelectedMonth : this.props.setDate('month'));
+			if ( !isDisabled )
+				props.onClick = ( this.props.updateOn === 'months' ?
+					this.updateSelectedMonth : this.props.setDate( 'month' ) );
 
-			months.push( renderer( props, i, year, date && date.clone() ));
+			months.push( renderer( props, i, year, date && date.clone() ) );
 
-			if ( months.length === 4 ){
-				rows.push( DOM.tr({ key: month + '_' + rows.length }, months) );
+			if ( months.length === 4 ) {
+				rows.push( DOM.tr({ key: month + '_' + rows.length }, months ) );
 				months = [];
 			}
 
@@ -66,26 +79,30 @@ var DateTimePickerMonths = React.createClass({
 	},
 
 	updateSelectedMonth: function( event ) {
-		this.props.updateSelectedDate(event, true);
+		this.props.updateSelectedDate( event );
 	},
 
 	renderMonth: function( props, month ) {
 		var localMoment = this.props.viewDate;
-		var monthStr = localMoment.localeData().monthsShort(localMoment.month(month));
+		var monthStr = localMoment.localeData().monthsShort( localMoment.month( month ) );
 		var strLength = 3;
 		// Because some months are up to 5 characters long, we want to
 		// use a fixed string length for consistency
-		var monthStrFixedLength = monthStr.substring(0, strLength);
+		var monthStrFixedLength = monthStr.substring( 0, strLength );
 		return DOM.td( props, capitalize( monthStrFixedLength ) );
 	},
 
-	isValidDate: function(){
+	alwaysValidDate: function() {
 		return 1;
-	}
-});
+	},
 
-function capitalize(str) {
-	return str.charAt(0).toUpperCase() + str.slice(1);
+  handleClickOutside: function() {
+    this.props.handleClickOutside();
+  }
+}));
+
+function capitalize( str ) {
+	return str.charAt( 0 ).toUpperCase() + str.slice( 1 );
 }
 
 module.exports = DateTimePickerMonths;