+++ src/YearsView.js
@@ -1,19 +1,22 @@
 'use strict';
 
-var React = require('react');
+var React = require('react'),
+    createClass = require('create-react-class'),
+	onClickOutside = require('react-onclickoutside')
+;
 
 var DOM = React.DOM;
-var DateTimePickerYears = React.createClass({
+var DateTimePickerYears = onClickOutside( createClass({
 	render: function() {
-		var year = parseInt(this.props.viewDate.year() / 10, 10) * 10;
+		var year = parseInt( this.props.viewDate.year() / 10, 10 ) * 10;
 
 		return DOM.div({ className: 'rdtYears' }, [
-			DOM.table({ key: 'a'}, DOM.thead({}, DOM.tr({}, [
-				DOM.th({ key: 'prev', className: 'rdtPrev' }, DOM.span({onClick: this.props.subtractTime(10, 'years')}, '‹')),
-				DOM.th({ key: 'year', className: 'rdtSwitch', onClick: this.props.showView('years'), colSpan: 2 }, year + '-' + (year + 9) ),
-				DOM.th({ key: 'next', className: 'rdtNext'}, DOM.span({onClick: this.props.addTime(10, 'years')}, '›'))
+			DOM.table({ key: 'a' }, DOM.thead({}, DOM.tr({}, [
+				DOM.th({ key: 'prev', className: 'rdtPrev', onClick: this.props.subtractTime( 10, 'years' )}, DOM.span({}, '‹' )),
+				DOM.th({ key: 'year', className: 'rdtSwitch', onClick: this.props.showView( 'years' ), colSpan: 2 }, year + '-' + ( year + 9 ) ),
+				DOM.th({ key: 'next', className: 'rdtNext', onClick: this.props.addTime( 10, 'years' )}, DOM.span({}, '›' ))
 				]))),
-			DOM.table({ key: 'years'}, DOM.tbody({}, this.renderYears( year )))
+			DOM.table({ key: 'years' }, DOM.tbody( {}, this.renderYears( year )))
 		]);
 	},
 
@@ -23,25 +26,37 @@ var DateTimePickerYears = React.createClass({
 			rows = [],
 			renderer = this.props.renderYear || this.renderYear,
 			selectedDate = this.props.selectedDate,
-			isValid = this.props.isValidDate || this.isValidDate,
-			classes, props
+			isValid = this.props.isValidDate || this.alwaysValidDate,
+			classes, props, currentYear, isDisabled, noOfDaysInYear, daysInYear, validDay,
+			// Month and date are irrelevant here because
+			// we're only interested in the year
+			irrelevantMonth = 0,
+			irrelevantDate = 1
 		;
 
 		year--;
-        var currentYear, disabled,
-            // Month and date are irrelevant here because
-            // we're only really interested in the year
-            irrelevantMonth = 1,
-            irrelevantDate = 1;
 		while (i < 11) {
 			classes = 'rdtYear';
 			currentYear = this.props.viewDate.clone().set(
-				{ year: year, month: irrelevantMonth, date: irrelevantDate });
-			if ( i === -1 | i === 10 )
-				classes += ' rdtOld';
+				{ year: year, month: irrelevantMonth, date: irrelevantDate } );
 
-			disabled = !isValid(currentYear);
-			if ( disabled )
+			// Not sure what 'rdtOld' is for, commenting out for now as it's not working properly
+			// if ( i === -1 | i === 10 )
+				// classes += ' rdtOld';
+
+			noOfDaysInYear = currentYear.endOf( 'year' ).format( 'DDD' );
+			daysInYear = Array.from({ length: noOfDaysInYear }, function( e, i ) {
+				return i + 1;
+			});
+
+			validDay = daysInYear.find(function( d ) {
+				var day = currentYear.clone().dayOfYear( d );
+				return isValid( day );
+			});
+
+			isDisabled = ( validDay === undefined );
+
+			if ( isDisabled )
 				classes += ' rdtDisabled';
 
 			if ( selectedDate && selectedDate.year() === year )
@@ -53,12 +68,13 @@ var DateTimePickerYears = React.createClass({
 				className: classes
 			};
 
-			if ( !disabled )
-				props.onClick = this.props.updateOn === 'years' ? this.updateSelectedYear : this.props.setDate('year');
+			if ( !isDisabled )
+				props.onClick = ( this.props.updateOn === 'years' ?
+					this.updateSelectedYear : this.props.setDate('year') );
 
 			years.push( renderer( props, year, selectedDate && selectedDate.clone() ));
 
-			if ( years.length === 4 ){
+			if ( years.length === 4 ) {
 				rows.push( DOM.tr({ key: i }, years ) );
 				years = [];
 			}
@@ -71,16 +87,20 @@ var DateTimePickerYears = React.createClass({
 	},
 
 	updateSelectedYear: function( event ) {
-		this.props.updateSelectedDate(event, true);
+		this.props.updateSelectedDate( event );
 	},
 
-	renderYear: function( props, year ){
+	renderYear: function( props, year ) {
 		return DOM.td( props, year );
 	},
 
-	isValidDate: function(){
+	alwaysValidDate: function() {
 		return 1;
-	}
-});
+	},
+
+  handleClickOutside: function() {
+    this.props.handleClickOutside();
+  }
+}));
 
 module.exports = DateTimePickerYears;